added sex information to published de novo datasets, and DDD dataset
This will allow us to identify probands with de novos in monoallelic
genes, and so exclude them from enrichment analyses (which improves
power for discovery of novel genes). We need to know the sex of the
removed probands, because we determine the expected mutation rates
using the numbers of male and female probands remaining in the cohort.