avoid error with pandas >=0.22.0 when all rates are missing
This fixes a bug where genes can gen p=0.0 when no rates data is available for a gene.
Since pandas 0.22.0, whenever summing a list where all values are NA, pandas will by
by default give sum=0.0, which if all gene rates are missing, ends up giving an
expected number of mutation=0, and this results in a P=0. This change reverts to the
previous pandas behaviour, where summing a list of NAs results in NA.